Change the way audio output object is created. Now, we do not destroy the audio output class object, instead we disable it. This allow to retrieve information about the audio capabilities such as the number of channels, if passthrough is available etc. The audio framework will now immediately return if it was fed rubbish and will not attempt to perform any initialisation. The greatly reduced the amount of open/close of the audio card.
Fixes #8792. Issue occured as an invalid codec detection lead to the audio framework not being initialised ; in turn it became impossible for AVFD to detect the capabilities of the hardware

git-svn-id: http://svn.mythtv.org/svn/trunk@26258 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/audioplayer.h
@@ -43,7 +43,7 @@ class MPUBLIC AudioPlayer
     bool  CanAC3(void);
     bool  CanDTS(void);
     uint  GetMaxChannels(void);
-    long long GetAudioTime(void);
+    int64_t GetAudioTime(void);
 
     bool      IsMuted(void) { return GetMuteState() == kMuteAll; }
     bool      SetMuted(bool mute);
